smartbank/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.7

# Redis and caching
redis==5.0.1
django-redis==5.4.0

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Encryption
cryptography==41.0.7

# File handling
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0
